refactor: move VecYawPitch out of BuilderBase
