+++ src/main/java/cam72cam/immersiverailroading/render/rail/RailRender.java
@@ -1,6 +1,7 @@
 package cam72cam.immersiverailroading.render.rail;
 
 import cam72cam.immersiverailroading.render.ExpireableMap;
+import cam72cam.immersiverailroading.render.VecYawPitch;
 import cam72cam.immersiverailroading.track.BuilderBase;
 import cam72cam.immersiverailroading.track.TrackBase;
 import cam72cam.immersiverailroading.util.RailInfo;
@@ -30,7 +31,7 @@ public class RailRender {
 	private final RailInfo info;
 	private boolean isLoaded;
 	private boolean isLoading;
-	private List<BuilderBase.VecYawPitch> renderData;
+	private List<VecYawPitch> renderData;
 	private List<TrackBase> tracks;
 
 	private RailRender(RailInfo info) {
