+++ src/main/java/cam72cam/immersiverailroading/track/BuilderCrossing.java
@@ -6,6 +6,7 @@ import java.util.List;
 import cam72cam.immersiverailroading.library.TrackModelPart;
 import cam72cam.immersiverailroading.util.RailInfo;
 import cam72cam.immersiverailroading.util.VecUtil;
+import cam72cam.immersiverailroading.render.VecYawPitch;
 import cam72cam.mod.math.Vec3d;
 import cam72cam.mod.math.Vec3i;
 import cam72cam.mod.world.World;
@@ -35,7 +36,7 @@ public class BuilderCrossing extends BuilderBase {
 	@Override
 	public List<VecYawPitch> getRenderData() {
 		
-		List<VecYawPitch> data = new ArrayList<VecYawPitch>();
+		List<VecYawPitch> data = new ArrayList<>();
 		float angle = 0;
 
 		double trackOffset = (1-info.settings.gauge.scale())/4;
