+++ src/main/java/cam72cam/immersiverailroading/track/BuilderCubicCurve.java
@@ -5,6 +5,7 @@ import cam72cam.immersiverailroading.library.TrackItems;
 import cam72cam.immersiverailroading.util.PlacementInfo;
 import cam72cam.immersiverailroading.util.RailInfo;
 import cam72cam.immersiverailroading.util.VecUtil;
+import cam72cam.immersiverailroading.render.VecYawPitch;
 import cam72cam.mod.item.ItemStack;
 import cam72cam.mod.math.Vec3d;
 import cam72cam.mod.math.Vec3i;
@@ -262,10 +263,7 @@ public class BuilderCubicCurve extends BuilderIterator {
 			List<VecYawPitch> data = new ArrayList<>();
 			for (BuilderBase curve : subBuilders.subList(0, Math.min(subBuilders.size(), 3))) {
 				Vec3d offset = new Vec3d(curve.pos.subtract(pos));
-				for (VecYawPitch rd : curve.getRenderData()) {
-					rd = new VecYawPitch(rd.x + offset.x, rd.y + offset.y, rd.z + offset.z, rd.yaw, rd.pitch, rd.length);
-					data.add(rd);
-				}
+				curve.getRenderData().stream().map(rd -> rd.add(offset)).forEach(data::add);
 			}
 			return data;
 		}
